requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
gprof2dot==2021.2.21
idna==2.10
Jinja2==2.11.3
loguru==0.5.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
pycodestyle==2.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
